Removing /orcid-grants/create from scopes
I was attempting to work through a failing automated test, and in an
attempt to exhaust solutions, when I receive the following error, I
figured I would add "/orcid-grants/create" to the list of
authentication scopes. The solution did not work, and based on a
response from support@orcid.org, they said that "/orcid-grants/create"
was an older scope that is no longer used.

ERROR
    <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
    <orcid-message xmlns="http://www.orcid.org/ns/orcid">
      <message-version>1.2_rc3</message-version>
      <error-desc>Bad Request : /orcid-grants/create</error-desc>
    </orcid-message>