Adding Orcid::NamedCallbacks
I've been thinking about teasing out an Orcid::Profile::Search runner.

* In one of the search context's finding a record would mean yielding
those records to the view to allow a user to pick one of the results.

* In another context finding one or more records would be reported to
the batch process that is handling multiple records (i.e. I found
these).

* In another context, when we have a user that we are going to attempt
to create an Orcid for, I want to have a guard clause that won't allow
you to create an Orcid for an existing email.