Adding error handling for "weird" search results
I don't know if these are truly weird retults, but the resulting
document from the query says there are 0 results but includes 10 nodes
that look like they could be results.

The returned JSON document is captured in the newly committed file:
    multiple-responses-without-valid-response.json

Given the potential volatility of the response document from Orcid,
I've also opted to extract the parser for the response document.
In doing so, it is easier to unit test parsing the response.

Closes #2

Relates to: ORCID/ORCID-Source#1100