Creating Orcid::ProfileStatus
This is an attempt to consolidate that knowledge.

There are several places in which we want to have access to a user's
Orcid Profile status, and methods related to those states.

The Orcid Profile Status makes use of callbacks because we do not
have a uniform object that represents each state for a user and their
orcid profile.

@TODO - review the controllers that have encapsulated this state and
have it lean on the Orcid::ProfileStatus handler.