Adding explicit pathing to partial rendering
In a reference implementation, we were encountering an error in which
the partials were not found. This is because the partial was
referenced as a file (i.e. filename) and not via a path (i.e.
path/to/filename). Rails has different lookup handling if given a
filename vs. a path.

http://guides.rubyonrails.org/layouts_and_rendering.html#using-partials