Adding more robust claim the orcid spec
Instead of relying on the very fragile UI, I'm leveraging an API
access to Mailinator. However, since the Sandbox is slow and
unpredictable in the timeliness of email delivery, I need to add a
retry.

The defaults in the mailinator requests method appear to work; But
things are still prone to breaking.