Improving user experience for profile creation
Prior to this commit, there was no confirmation that the profile
request was. The process was also injecting confusing redirects.

Many of the routing decisions in this micro-application are not
adequate for implementing applications.

By introducing the #path_for method, I'm hoping to provide a means
for adopters to override redirect behavior.

Closes #41