Removing query attributes from ProfileConnection
Having more than one field is confusing, in that it is treated as an
AND condition. The :text field is a superset of both the DOI and the
email field, and provides a more powerful search option.

This is of note in that many people will hide their email address by
default.